requirements.txt
aiofiles==24.1.0
aiogram==3.14.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiosmtplib==3.0.2
annotated-types==0.7.0
asyncpg==0.30.0
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
et_xmlfile==2.0.0
fake-useragent==1.5.1
fonttools==4.55.0
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
jpype1==1.5.1
kiwisolver==1.4.7
logging==0.4.9.6
magic-filter==1.0.12
multidict==6.1.0
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pdf2image==1.17.0
pillow==11.0.0
plum-dispatch==1.7.4
propcache==0.2.0
psycopg2==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
pylovepdf==1.3.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.17.1
